+++ src/components/newScheduleForm/DatePicker.jsx
@@ -1,15 +1,14 @@
-import { DatePicker } from "antd";
+import { DatePicker as AntdDatePicker } from "antd";
 
-//TODO: 시간 따로 날짜 따로 입력하기
 //TODO: 키보드 입력 거름망
 
-export default function DateInput({ onChange, ...props }) {
+export default function DatePicker({ onChange, ...props }) {
     //TODO: Form과 연결(param 설정)
     //TODO: Label 클릭시 다시 열리는 현상 수정
     //TODO: enter 입력시 제출하는 현상 수정
 
     return (
-        <DatePicker
+        <AntdDatePicker
             format="YY년 MM월 DD일"
             onChange={(newValue) => {
                 onChange(newValue);
