+++ src/components/newScheduleForm/index.js
@@ -1,6 +1,6 @@
 import NewScheduleForm from "./NewScheduleForm";
 
-export { default as DateInput } from "./DateInput";
+export { default as DatePicker } from "./DatePicker";
 export { default as OptionLabel } from "./OptionLabel";
-export { default as TimeInput } from "./TimeInput";
+export { default as TimePicker } from "./TimePicker";
 export default NewScheduleForm;
